Use Quarto navbar to set the CSS
To see a demo, open the page, delete the navbar created by quarto.
(This should leave only the html navbar that is included). Now it
looks ok, except that the header is disapearing. Turn off the
position: fixed attribute in the css, then see that it is not at
the top of the page. I'm not sure what is going on here.